requirements.txt
fastapi
uvicorn
pymongo
psycopg2-binary
neo4j
pyspark
python-dotenv